move main cli stuff to separate file
